requirements.txt
pyrogram==2.0.39
python-dotenv==1.0.0